+++ processing_unit.go
@@ -70,7 +70,7 @@ func (pu *ProcessingUnit) Get(desc descriptors.ProcessingUnitControlDescriptor) 
 	ifnum := pu.usb.altsetting.bInterfaceNumber
 
 	bufLen := 16
-	buf := C.malloc(C.ulong(bufLen))
+	buf := C.malloc(C.size_t(bufLen))
 	defer C.free(buf)
 
 	if ret := C.libusb_control_transfer(
